refactor: :recycle: Refactor method calls
